requirements.txt
openai
slack_sdk
python-dotenv
pyyaml
